Refactor info requests to a special class
